+++ ModelModel/Base.lproj/Main.storyboard
@@ -9,10 +9,10 @@
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
     <scenes>
-        <!--First Table View Controller-->
+        <!--User Name Table View Controller-->
         <scene sceneID="6HK-W9-516">
             <objects>
-                <tableViewController id="VE3-gK-kH4" customClass="FirstTableViewController" customModule="ModelModel" customModuleProvider="target" sceneMemberID="viewController">
+                <tableViewController id="VE3-gK-kH4" customClass="UserNameTableViewController" customModule="ModelModel" customModuleProvider="target" sceneMemberID="viewController">
                     <tableView key="view" clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" dataMode="prototypes" style="plain" separatorStyle="default" rowHeight="44" sectionHeaderHeight="28" sectionFooterHeight="28" id="mJD-PJ-EDz">
                         <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
@@ -50,10 +50,10 @@
             </objects>
             <point key="canvasLocation" x="879" y="152"/>
         </scene>
-        <!--Second Table View Controller-->
+        <!--User Characteristics Table View Controller-->
         <scene sceneID="Fd8-73-G9e">
             <objects>
-                <tableViewController id="vAO-1u-m9m" customClass="SecondTableViewController" customModule="ModelModel" customModuleProvider="target" sceneMemberID="viewController">
+                <tableViewController id="vAO-1u-m9m" customClass="UserCharacteristicsTableViewController" customModule="ModelModel" customModuleProvider="target" sceneMemberID="viewController">
                     <tableView key="view" clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" dataMode="prototypes" style="plain" separatorStyle="default" rowHeight="44" sectionHeaderHeight="28" sectionFooterHeight="28" id="vmr-XZ-vtB">
                         <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
